requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
article==0.1.1
astor==0.8.1
async-generator==1.10
attrs==20.3.0
b64==0.4
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.0
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.7.1
bleach==3.3.0
blinker==1.4
blis==0.7.4
bokeh==2.3.0
breadability==0.1.20
bs4==0.0.1
cached-properties==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
colorama==0.4.4
comparative==0.0.2
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
datasketch==1.5.3
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
en==0.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
fish==1.1
future==0.18.2
fuzz==0.1.1
fuzzy-sequence-matcher==0.1.1
fuzzywuzzy==0.18.0
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.14
GoogleNews==1.5.5
idna==2.10
inflect==5.3.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
jedi==0.18.0
jellyfish==0.8.2
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.7.0
MouseInfo==0.1.3
murmurhash==1.0.5
mutagen==1.45.1
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
newsapi==0.1.1
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
notebook==6.2.0
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
Pattern==3.6
pdfminer.six==20201018
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
pluralize==0.1.6
portend==2.7.1
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.15.6
pyarrow==3.0.0
PyAutoGUI==0.9.52
pycountry==20.7.3
pycparser==2.20
pydantic==1.7.3
pydeck==0.6.1
PyGetWindow==0.0.9
Pygments==2.8.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
Pyphen==0.10.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
python-docx==0.8.10
python-Levenshtein-wheels==0.13.2
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
retrying==1.3.3
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
smart-open==3.0.0
smmap==3.0.5
sortedcontainers==2.3.0
soupsieve==2.2
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
stompest==2.3.0
streamlit==0.78.0
strsim==0.0.3
summarizer==0.0.7
sumy==0.8.1
tempora==4.0.1
terminado==0.9.3
testpath==0.4.4
textblob==0.15.3
textstat==0.7.0
thinc==8.0.2
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.1
toml==0.10.2
toolz==0.11.1
torch==1.8.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.4.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vaderSentiment==3.3.2
validators==0.18.2
wasabi==0.8.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wxPython==4.1.1
xlrd==2.0.1
zc.lockfile==2.0
